+++ ibis/backends/mysql/__init__.py
@@ -1,4 +1,4 @@
-from ibis.sql.alchemy import to_sqlalchemy
+from ibis.backends.base_sqlalchemy.alchemy import to_sqlalchemy
 
 from .client import MySQLClient
 from .compiler import dialect, rewrites  # noqa: F401
