+++ ibis/backends/mysql/client.py
@@ -7,7 +7,7 @@ import sqlalchemy as sa
 import sqlalchemy.dialects.mysql as mysql
 
 import ibis.expr.datatypes as dt
-import ibis.sql.alchemy as alch
+import ibis.backends.base_sqlalchemy.alchemy as alch
 
 from .compiler import MySQLDialect
