+++ ibis/backends/sqlite/__init__.py
@@ -21,7 +21,7 @@ def compile(expr, params=None):
     """
     Force compilation of expression for the SQLite target
     """
-    from ibis.sql.alchemy import to_sqlalchemy
+    from ibis.backends.base_sqlalchemy.alchemy import to_sqlalchemy
 
     return to_sqlalchemy(expr, dialect.make_context(params=params))
